django_test/requirements.txt
Django==1.5
defusedxml==0.4.1
django-tastypie==0.10.0
lxml==3.2.3
mimeparse==0.1.3
python-dateutil==2.1
six==1.4.1
